chore: inline Noop tracer/span singletons + delete dead RegistryFunctions methods
Two mechanical cleanups from the #165 architecture audit:

1. Collapse `NoopTracer` / `NoopSpanScope` into nested enum singletons inside
   `Tracer` / `Tracer.SpanScope`. The two package-private classes existed only
   as `INSTANCE` holders reachable through the `Tracer.noop()` /
   `SpanScope.noop()` factories — the idiomatic Java enum-singleton form
   eliminates the two files without any callsite change.

2. Delete `RegistryFunctions.createMetrics` and
   `RegistryFunctions.withFunctionErrorHandling` — both are unreferenced in
   production code (only `RegistryFunctionsTest` called them). Remove the
   corresponding tests, fix the class-level Javadoc reference to a
   non-existent `withErrorHandling(fn, defaultValue)` overload, and keep
   `withConsumerErrorHandling` / `withOperatorErrorHandling` /
   `createUnmodifiableView` which `MessageProcessorRegistry` still uses.

Net: -124 lines, zero behavior change (factory seams preserved, deleted
helpers had no production callers).

Closes #165
