--- lib/kpipe-producer/src/main/java/io/github/eschizoid/kpipe/producer/tracing/NoopSpanScope.java
@@ -1,20 +1,0 @@
-package io.github.eschizoid.kpipe.producer.tracing;
-
-/// Zero-cost default [Tracer.SpanScope] returned by [NoopTracer]. Package-private — callers reach
-/// it through [Tracer.SpanScope#noop()].
-final class NoopSpanScope implements Tracer.SpanScope {
-
-  static final NoopSpanScope INSTANCE = new NoopSpanScope();
-
-  private NoopSpanScope() {}
-
-  @Override
-  public void recordException(final Throwable t) {
-    // no-op
-  }
-
-  @Override
-  public void close() {
-    // no-op
-  }
-}
