+++ lib/kpipe-producer/src/main/java/io/github/eschizoid/kpipe/producer/tracing/package-info.java
@@ -2,7 +2,8 @@
 ///
 /// - [Tracer] — the SPI implemented by tracing backends; produces span scopes around
 ///   producer sends and injects propagation headers (e.g. W3C `traceparent`) into Kafka records.
-/// - [NoopTracer] / [NoopSpanScope] — zero-overhead defaults used when no backend is wired in.
+/// - [Tracer#noop()] / [Tracer.SpanScope#noop()] — zero-overhead defaults used when no backend
+///   is wired in.
 ///
 /// An OpenTelemetry-backed implementation is provided by `kpipe-tracing-otel`. The library
 /// code in `kpipe-consumer` / `kpipe-producer` depends only on this SPI and has no transitive
